requirements.txt
Flask==2.1.2
gunicorn==20.1.0
ibm_cloud_sdk_core==3.15.3
ibm_cos_sdk==2.11.0
ibm_cos_sdk_core==2.11.0
ibm_watson==6.0.0
ibmcloudant==0.1.4
python-dotenv==0.21.0
requests==2.28.1
twilio==7.1.0
Werkzeug==2.0.1